requirements.txt
numpy>=1.20.0
pandas>=1.3.0
pyarrow>=5.0.0
click>=8.0.0
pyyaml>=5.4.0
tushare>=1.2.60
pyqlib>=0.8.0
scikit-learn>=1.0.0
xgboost>=1.5.0
lightgbm>=3.3.0
catboost>=1.0.0
torch>=1.10.0
sqlalchemy>=1.4.0
requests>=2.26.0
loguru>=0.5.3
plotly>=5.0.0
